Abstract Unity error log intercepting out of TestInstructionExecutor
