+++ Packages/Responsible/Runtime/TestInstructionExecutor.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
 using JetBrains.Annotations;
@@ -10,7 +9,6 @@ using Responsible.Context;
 using Responsible.State;
 using Responsible.Utilities;
 using UnityEngine;
-using UnityEngine.TestTools;
 
 namespace Responsible
 {
@@ -30,9 +28,9 @@ namespace Responsible
 			new SafeIterationList<Action<TestOperationStateNotification>>();
 
 		private readonly CancellationTokenSource mainCancellationTokenSource = new CancellationTokenSource();
-		private readonly List<(LogType type, Regex regex)> expectedLogs = new List<(LogType, Regex)>();
-		private readonly ILogger logger;
 		private readonly ITimeProvider timeProvider;
+		private readonly IExternalResultSource externalResultSource;
+		private readonly ILogger logger;
 
 		/// <summary>
 		/// Will call <paramref name="callback"/> when operations start or finish.
@@ -55,9 +53,11 @@ namespace Responsible
 		/// </param>
 		public TestInstructionExecutor(
 			ITimeProvider timeProvider,
+			IExternalResultSource externalResultSource = null,
 			ILogger logger = null)
 		{
 			this.timeProvider = timeProvider;
+			this.externalResultSource = externalResultSource;
 			this.logger = logger ?? Debug.unityLogger;
 		}
 
@@ -71,22 +71,6 @@ namespace Responsible
 			// leave e.g. static Unity log event subscriptions hanging.
 			this.mainCancellationTokenSource.Cancel();
 			this.mainCancellationTokenSource.Dispose();
-		}
-
-		/// <summary>
-		/// Expect a log message during execution of any subsequent test instructions.
-		/// Similar to <see cref="LogAssert.Expect(LogType, Regex)"/>,
-		/// but needs to be called instead of that, in order to work when executing test instructions.
-		/// </summary>
-		/// <remarks>
-		/// Will call <see cref="LogAssert.Expect(LogType, Regex)"/>, so it doesn't need to be called separately.
-		/// </remarks>
-		/// <param name="logType">Log type to expect.</param>
-		/// <param name="regex">Regex to match the message to be expected.</param>
-		public void ExpectLog(LogType logType, Regex regex)
-		{
-			LogAssert.Expect(logType, regex);
-			this.expectedLogs.Add((logType, regex));
 		}
 
 		internal async Task<T> RunInstruction<T>(
@@ -103,9 +87,16 @@ namespace Responsible
 					NotificationCallbacks.ForEach(cb =>
 						cb(new TestOperationStateNotification.Started(rootState)));
 
-					return await linkedTokenSource.Token.Amb(
-						this.InterceptErrors<T>,
-						ct => rootState.Execute(runContext, ct));
+					if (this.externalResultSource != null)
+					{
+						return await linkedTokenSource.Token.Amb(
+							this.externalResultSource.GetExternalResult<T>,
+							ct => rootState.Execute(runContext, ct));
+					}
+					else
+					{
+						return await rootState.Execute(runContext, linkedTokenSource.Token);
+					}
 				}
 				catch (Exception e)
 				{
@@ -151,51 +142,5 @@ namespace Responsible
 				$"Test operation execution {what}!",
 				$"Failure context:\n{StateStringBuilder.MakeState(rootOperation)}",
 			};
-
-		// Intercept errors respecting and not toggling the globally mutable (yuck) LogAssert.ignoreFailingMessages
-		private async Task<T> InterceptErrors<T>(CancellationToken token)
-		{
-			var completionSource = new TaskCompletionSource<T>();
-
-			void HandleSingleLog(string condition, LogType type)
-			{
-				var index = this.expectedLogs.FindIndex(entry =>
-					entry.type == type &&
-					entry.regex.IsMatch(condition));
-
-				if (index >= 0)
-				{
-					// Already expected, just remove it
-					this.expectedLogs.RemoveAt(index);
-				}
-				else
-				{
-					LogAssert.Expect(type, condition);
-					completionSource.SetException(new UnhandledLogMessageException(condition));
-				}
-			}
-
-			void LogHandler(string condition, string _, LogType type)
-			{
-				if (!LogAssert.ignoreFailingMessages &&
-					(type == LogType.Error || type == LogType.Exception))
-				{
-					HandleSingleLog(condition, type);
-				}
-			}
-
-			using (token.Register(() => completionSource.SetCanceled()))
-			{
-				Application.logMessageReceived += LogHandler;
-				try
-				{
-					return await completionSource.Task;
-				}
-				finally
-				{
-					Application.logMessageReceived -= LogHandler;
-				}
-			}
-		}
 	}
 }
