+++ Packages/Responsible/Runtime/Unity/UnityErrorLogInterceptor.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 324cf5fc4ee84c6cab6d7520dbeb7b5d
+timeCreated: 1616448028
